Doc:Replace plugin_header file with plugin_header-integration file
Integration plugins need a different header from stand-alone plugins.
The plugin docs should point to the integration repo rather than the
stand-alone input and output repo. We also add a note that the plugin
is part of an integration rather than stand-alone.

The new header included in this work was added with Logstash #11891.