Improve error handling
This commit improves the error handling when polling and commiting offsets by:

Adding begin...rescue blocks around the poll and commitSync methods in the polling thread
to avoid handlable errors escaping out of the plugin, and causing the plugin, or even the full Logstash
process, to crash.

Use Java threads instead of Ruby threads to enable the default uncaught exception handler to correctly
handle uncaught exceptions in the polling thread. Without this, these errors are "handled" according to
`Thread.abort_on_exception` is configured; if this is set to true - as it currently is on `master` and
`7.x`, Logstash execution will stop immediately across all pipelines. Setting this value to false has other
side effects - OutOfMemoryErrors or other fatal errors will not be caught by the default uncaught exception handler
and execution will continue, leading to undefined behaviour.