Add option to skip validation when dealing with schema registry (#97)
* Add option to skip vaildation when dealing with schema registry

This PR adds a `schema_registry_validation` option to the plugin, allowing
users to bypass the schema registry validation while using a plugin. This is
a common issue when using authenticated schema registry, as the validation occurs
outside of the 'KafkaConsumer' code, and will require considerable effort to maintain
parity with that method.

This configuration option has 2 settings `auto` and `skip`. The default `auto` setting
will skip validation when schema registry auth is used that is known to fail, currently
when Kerberos authentication is used, but will try and validate in other cases. `skip`
will skip regardless of authentication method.

Co-authored-by: Ry Biesemeyer <yaauie@users.noreply.github.com>
Co-authored-by: Karen Metts <35154725+karenzone@users.noreply.github.com>