AWS IAM authentication (#178)
Creation and use of new setting to load uber jars which contains cloud providers IAM authentication SASL client libraries.

Permit to the users that deploy Logstash in a cloud provider to leverage the cloud provider's IAM to authenticate the client with SASL.

Introduces the setting `sasl_iam_jar_paths` which is an array of library paths that points to uber jars containing all the classes needed by specific cloud provider IAM for SASL authentication.
During registration phase, uses the Ruby require to load the jars configured.

---------

Co-authored-by: Akos Korsos <akos.korsos@n26.com>
Co-authored-by: João Duarte <jsvd@users.noreply.github.com>